requirements.txt
kafka-python>=2.0.1

